Рефакторинг Питона.
