+++ server/weapons/python.cpp
@@ -13,14 +13,7 @@
 *
 ****/
 
-#include "extdll.h"
-#include "util.h"
-#include "cbase.h"
-#include "weapons.h"
-#include "monsters.h"
-#include "player.h"
-#include "gamerules.h"
-#include "user_messages.h"
+#include "python.h"
 
 enum python_e
 {
@@ -32,27 +25,6 @@ enum python_e
 	PYTHON_DRAW,
 	PYTHON_IDLE2,
 	PYTHON_IDLE3
-};
-
-class CPython : public CBasePlayerWeapon
-{
-	DECLARE_CLASS( CPython, CBasePlayerWeapon );
-public:
-	void Spawn( void );
-	void Precache( void );
-	int iItemSlot( void ) { return 2; }
-	int GetItemInfo(ItemInfo *p);
-	int AddToPlayer( CBasePlayer *pPlayer );
-	void PrimaryAttack( void );
-	void SecondaryAttack( void );
-	BOOL Deploy( void );
-	void Holster( void );
-	void Reload( void );
-	void WeaponIdle( void );
-	float m_flSoundDelay;
-
-	BOOL m_fInZoom;// don't save this. 
-	int m_iShell;
 };
 
 LINK_ENTITY_TO_CLASS( weapon_python, CPython );
@@ -293,31 +265,3 @@ void CPython::WeaponIdle( void )
 	}
 	SendWeaponAnim( iAnim );
 }
-
-class CPythonAmmo : public CBasePlayerAmmo
-{
-	DECLARE_CLASS( CPythonAmmo, CBasePlayerAmmo );
-
-	void Spawn( void )
-	{ 
-		Precache( );
-		SET_MODEL(ENT(pev), "models/w_357ammobox.mdl");
-		CBasePlayerAmmo::Spawn( );
-	}
-	void Precache( void )
-	{
-		PRECACHE_MODEL ("models/w_357ammobox.mdl");
-		PRECACHE_SOUND("items/9mmclip1.wav");
-	}
-	BOOL AddAmmo( CBaseEntity *pOther ) 
-	{ 
-		if (pOther->GiveAmmo( AMMO_357BOX_GIVE, "357", _357_MAX_CARRY ) != -1)
-		{
-			EMIT_SOUND(ENT(pev), CHAN_ITEM, "items/9mmclip1.wav", 1, ATTN_NORM);
-			return TRUE;
-		}
-		return FALSE;
-	}
-};
-
-LINK_ENTITY_TO_CLASS( ammo_357, CPythonAmmo );
