+++ server/weapons/python.h
@@ -1,0 +1,46 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*
+*	This product contains software technology licensed from Id
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc.
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+
+#pragma once
+
+#include "extdll.h"
+#include "util.h"
+#include "cbase.h"
+#include "weapons.h"
+#include "monsters.h"
+#include "player.h"
+#include "gamerules.h"
+#include "user_messages.h"
+
+class CPython : public CBasePlayerWeapon
+{
+	DECLARE_CLASS(CPython, CBasePlayerWeapon);
+public:
+	void Spawn(void);
+	void Precache(void);
+	int iItemSlot(void) { return 2; }
+	int GetItemInfo(ItemInfo *p);
+	int AddToPlayer(CBasePlayer *pPlayer);
+	void PrimaryAttack(void);
+	void SecondaryAttack(void);
+	BOOL Deploy(void);
+	void Holster(void);
+	void Reload(void);
+	void WeaponIdle(void);
+	float m_flSoundDelay;
+
+	BOOL m_fInZoom;// don't save this. 
+	int m_iShell;
+};
